perf: monaco 懒加载 + CI 工作流
monaco 懒加载:
- CodeEditor/CodeViewer 拆出 Impl 文件, index.tsx 改为 React.lazy 封装
- 消费方零改动, Suspense 内置 fallback=Spin
- 启用 esbuildMinifyIIFE 解决 helper 冲突
- CodeEditor/CodeViewer 仅在实际渲染时加载 monaco chunk

CI 工作流:
- 新增 .github/workflows/ci.yml
- PR/push 自动执行 install + typecheck + build + test
