+++ src/components/CodeViewer/index.tsx
@@ -1,57 +1,24 @@
-import React from 'react';
-import MonacoEditor from './MonacoEditor';
-import styles from './index.less';
+import { Spin } from 'antd';
+import React, { Suspense } from 'react';
 
-interface CodeViewerProps {
-  /** 是否动态主题 */
-  isDynamicTheme?: boolean;
-  /** 文件ID */
-  fileId: string;
-  /** 文件名 */
-  fileName: string;
-  /** 文件路径 */
-  filePath: string;
-  /** 文件内容 */
-  content: string;
-  /** 是否只读 */
-  readOnly: boolean;
-  /** 内容变化回调 */
-  onContentChange: (fileId: string, content: string) => void;
-}
+import type { CodeViewerProps } from './CodeViewerImpl';
 
-/**
- * 代码查看器组件
- * 显示代码内容，使用 MonacoEditor
- */
-const CodeViewer: React.FC<CodeViewerProps> = ({
-  isDynamicTheme = false,
-  fileId,
-  fileName,
-  filePath,
-  content,
-  readOnly,
-  onContentChange,
-}) => {
+const LazyCodeViewer = React.lazy(
+  () => import(/* webpackChunkName: "monaco-bundle" */ './CodeViewerImpl'),
+);
+
+function CodeViewer(props: CodeViewerProps) {
   return (
-    <div className={styles.fileContentDisplay}>
-      <MonacoEditor
-        key={fileId}
-        isDynamicTheme={isDynamicTheme}
-        currentFile={{
-          id: fileId,
-          name: fileName,
-          type: 'file',
-          path: filePath,
-          content: content,
-          lastModified: Date.now(),
-          children: [],
-        }}
-        onContentChange={onContentChange}
-        readOnly={readOnly}
-        className={styles.monacoEditor}
-      />
-    </div>
+    <Suspense
+      fallback={
+        <div style={{ display: 'flex', justifyContent: 'center', padding: 24 }}>
+          <Spin />
+        </div>
+      }
+    >
+      <LazyCodeViewer {...props} />
+    </Suspense>
   );
-};
+}
 
 export default CodeViewer;
